Move the tab validity check
Previously there was a bug when chrome first started as the tab
history was from the previous session so all of the tab ids were
invalid. To avoid this issue we move the tab existence check to the
chrome.commands.onCommand.addListener() and since we no longer need it
in the chrome.tabs.onActivated.addListener() we remove it from
there. The chrome.tabs.onActivated.addListener() can simply add the
newly activated tab to the history, without worrying about the
validity of the other tabs in the history.