Refactored background.js
ChatGPT was asked: Can you see any other ways to simplify the code and
reduce the number of times we need to access the chrome.tabs API?

Here are a few more improvements we can make:

1. Currently, the switchTab and switchBetweenCurrentAndLast functions
are fetching the tabIdHistory from the storage, even though we're
already doing that in the chrome.commands.onCommand.addListener(). We
can modify these functions to accept the tabIdHistory as a parameter
to avoid fetching it twice.

2. Similarly, the switchTab function is saving the tabIdHistory to
storage immediately after switching tabs. We can also do this in the
chrome.commands.onCommand.addListener(), after the command has been
executed.

3. It seems that the array.splice() function within the for loop in
chrome.commands.onCommand.addListener() may lead to an incorrect
operation as it changes the array during the iteration. It might be
better to use a filter function to clean the history.