Add extended history based on below request
One issue is where the user finds they have a lot of open tabs and they decide to shut down some of
them. When they do this they can end up closing the same number or more tabs than currently being
remembered. This has the effect of making the tab switcher non-functional until more tabs have been
visited once again. It seems that perhaps the tab switcher could keep a record of all of the tabs
that have been opened. It would only cycle through the numTabs, but where one of those tabs is
closed by the user it can add the most recently used tab (outside of numTabs) to the list of tabs to
cycle through